refactor: move some testing from `db/mod.rs` to integration testing
